:sparkles: Feat: 동행 신청 구현
- application db 생성
- 동행 신청 구현

Related to: #24
