--- src/main/java/com/example/festimo/admin/dto/AdminDTO.java
@@ -1,22 +1,0 @@
-package com.example.festimo.admin.dto;
-
-import java.time.LocalDateTime;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@AllArgsConstructor
-@Getter
-@Setter
-public class AdminDTO {
-    private Long userId;
-    private String userName;
-    private String nickname;
-    private String email;
-    private String role; // ADMIN, USER
-    private LocalDateTime createdDate;
-    private String gender; // F,M
-    private Float ratingAvg;
-
-}
