+++ src/main/java/com/example/festimo/domain/admin/repository/CommunityPostsRepository.java
@@ -1,0 +1,10 @@
+package com.example.festimo.domain.admin.repository;
+
+
+import com.example.festimo.domain.admin.entity.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CommunityPostsRepository extends JpaRepository<Post, Long> {
+}
