+++ src/main/java/com/example/festimo/domain/meet/dto/ApplicationRequest.java
@@ -1,0 +1,17 @@
+package com.example.festimo.domain.meet.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ApplicationRequest {
+
+    private Long userId;
+    private Long companyId;
+
+}
