+++ src/main/java/com/example/festimo/domain/meet/dto/CompanionRequest.java
@@ -1,0 +1,13 @@
+package com.example.festimo.domain.meet.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class CompanionRequest {
+    private Long userId;
+    private Long companyId;
+}
