+++ src/main/java/com/example/festimo/domain/meet/entity/Applications.java
@@ -1,0 +1,41 @@
+package com.example.festimo.domain.meet.entity;
+
+import java.time.LocalDateTime;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor
+@Getter
+public class Applications {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long applicationId;
+
+    @Column( nullable = false)
+    private Long userId;
+
+    @Column( nullable = false)
+    private Long companyId;
+
+    @Column(nullable = false)
+    private Status status = Status.PENDING; // 기본값은 'PENDING'
+
+    @Column(nullable = false)
+    private LocalDateTime appliedDate = LocalDateTime.now();
+
+    public Applications(Long userId, Long companyId) {
+        this.userId = userId;
+        this.companyId = companyId;
+    }
+
+    public enum Status {
+        PENDING,
+        APPROVED,
+        REJECTED
+    }
+
+}
