+++ src/main/java/com/example/festimo/domain/meet/entity/Companies.java
@@ -1,0 +1,18 @@
+package com.example.festimo.domain.meet.entity;
+
+import java.time.LocalDateTime;
+
+import jakarta.persistence.*;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor
+public class Companies {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long companyId;
+
+    @Column(nullable = false)
+    private LocalDateTime companyDate;
+}
