+++ src/main/java/com/example/festimo/domain/meet/entity/Companions.java
@@ -1,0 +1,20 @@
+package com.example.festimo.domain.meet.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class Companions {
+
+    @EmbeddedId
+    private CompanionId id;
+
+
+}
