+++ src/main/java/com/example/festimo/domain/user/repository/UserRepository.java
@@ -1,4 +1,12 @@
 package com.example.festimo.domain.user.repository;
 
-public interface UserRepository {
+import com.example.festimo.domain.user.entity.Users;
+import lombok.NonNull;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<Users,Long> {
+    @NonNull
+    Page<Users> findAll(@NonNull Pageable pageable);
 }
